client: refactor post allocrunnerv2 finalization
